Use consistent implementation of matmul in puzzles 16, 19 and 22. (#84)
Both puzzles 19 and 22 require an updated version of puzzle 16 to support different layouts for each of the input (a and b) and output LayoutTensors.
Additionaly added explicit checks for out of bound input and output elements which are not strictly necessary because copy_dram_to_sram_async fills out of bound elements with zeros.